feat(iframe): uniformize /iframe and /editiframe & add /bazariframe to distinguish use for widgets & add header/footer for login prompt in /show
