+++ tools/bazar/presentation/templates/liste_liens.tpl.html
@@ -35,6 +35,4 @@ if(count($fiches)>0) : ?>
     endforeach; ?>
     </ul>
     <?php echo $pager_links;?>
-    <?php
-endif;
-<?php echo $pager_links;?>
+<?php endif; ?>
