+++ tools/templates/handlers/page/show__.php
@@ -25,17 +25,13 @@ if (isset($GLOBALS['template-error']) && $GLOBALS['template-error']['type'] == '
 }
 
 if (!$this->HasAccess('read')) {
-    $output = '';
-    // on recupere les entetes html mais pas ce qu'il y a dans le body
-    $header = explode('<body', $this->Header());
-    $output .= $header[0] . '<body class="login-body">'."\n"
+    $output = $this->Header();
+    $output .= '<body class="login-body">'."\n"
         .'<div class="container">'."\n"
         .'<div class="yeswiki-page-widget page-widget page" '.$this->Format('{{doubleclic iframe="1"}}').'>'."\n";
 
     if ($contenu = $this->LoadPage("PageLogin")) {
         $output .= $this->Format($contenu["body"]);
-        // on recupere juste les javascripts et la fin des balises body et html
-        $output .= preg_replace('/^.+<script/Us', '<script', $this->Footer());
     } else {
         // sinon on affiche le formulaire d'identification minimal
         $output .= str_replace(
@@ -47,8 +43,10 @@ if (!$this->HasAccess('read')) {
             $plugin_output_new
         );
     }
-    // on recupere juste les javascripts et la fin des balises body et html
-    $output .= preg_replace('/^.+<script/Us', '<script', $this->Footer());
+    // common footer for all iframe page
+    $output .= '</div><!-- end .page-widget -->' . "\n";
+    $output .= '</div><!-- end .container -->' . "\n";
+    $output .= $this->Footer();
     exit($output);
 }
