+++ tools/templates/handlers/page/widget.php
@@ -11,8 +11,8 @@ $height = (isset($_GET['height'])) ? $_GET['height'] : 700;
 echo $this->Header();
 echo "<h2>"._t('TEMPLATE_WIDGET_TITLE')."</h2>";
 echo "<pre>\n";
-echo htmlentities('<iframe class="yeswiki_frame" width="'.$width.'" height="'.$height.'" frameborder="0" src="'.$this->Href('iframe').'"></iframe>')."\n";
+echo htmlentities('<iframe class="yeswiki_frame" width="'.$width.'" height="'.$height.'" frameborder="0" src="'.$this->Href('bazariframe').'"></iframe>')."\n";
 echo '</pre>'."\n";
 echo '<div class="alert alert-info">'._t('TEMPLATE_WIDGET_COPY_PASTE').'</div>'."\n";
-echo '<iframe class="yeswiki_frame" width="'.$width.'" height="'.$height.'" frameborder="0" src="'.$this->Href('iframe').'"></iframe>';
+echo '<iframe class="yeswiki_frame" width="'.$width.'" height="'.$height.'" frameborder="0"></iframe>';
 echo $this->Footer();
